requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
pymongo
python-dotenv
weasyprint
WTForms
mysql-connector-python